+++ tests/Unit/Colors/Hsl/ChannelTest.php
@@ -4,12 +4,14 @@ declare(strict_types=1);
 
 namespace Intervention\Image\Tests\Unit\Colors\Hsl;
 
+use Generator;
 use PHPUnit\Framework\Attributes\CoversClass;
 use Intervention\Image\Colors\Hsl\Channels\Hue;
 use Intervention\Image\Colors\Hsl\Channels\Saturation;
 use Intervention\Image\Colors\Hsl\Channels\Luminance;
 use Intervention\Image\Exceptions\InvalidArgumentException;
 use Intervention\Image\Tests\BaseTestCase;
+use PHPUnit\Framework\Attributes\DataProvider;
 
 #[CoversClass(Hue::class)]
 #[CoversClass(Saturation::class)]
@@ -84,5 +86,34 @@ final class ChannelTest extends BaseTestCase
 
         $this->expectException(InvalidArgumentException::class);
         new Luminance(-1);
+    }
+
+    #[DataProvider('scaleDataProvider')]
+    public function testScale(int $value, int $percent, int $result): void
+    {
+        $this->assertEquals($result, (new Hue($value))->scale($percent)->value());
+    }
+
+    public static function scaleDataProvider(): Generator
+    {
+        yield [0, 0, 0];
+        yield [360, 0, 360];
+        yield [180, 0, 180];
+
+        yield [0, 50, 180];
+        yield [360, 50, 360];
+        yield [180, 50, 270];
+
+        yield [0, 100, 360];
+        yield [360, 100, 360];
+        yield [180, 100, 360];
+
+        yield [0, -50, 0];
+        yield [360, -50, 180];
+        yield [180, -50, 90];
+
+        yield [0, -100, 0];
+        yield [360, -100, 0];
+        yield [180, -100, 0];
     }
 }