+++ changelog.md
@@ -88,6 +88,7 @@
 - ImageInterface::drawEllipse() has a different signature: coordinate arguments removed
 - ImageInterface::drawRectangle() has a different signature: coordinate arguments removed
 - Added ColorInterface::withTransparency()
+- Added ColorChannelInterface::scale()
 
 ### Exceptions